requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.115.0
frozenlist==1.4.1
google-ai-generativelanguage==0.6.6
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.2
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
ibm-cos-sdk==2.13.6
ibm-cos-sdk-core==2.13.6
ibm-cos-sdk-s3transfer==2.13.6
ibm_watsonx_ai==1.1.11
idna==3.10
importlib_metadata==8.5.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.1
langchain-core==0.3.6
langchain-google-genai==2.0.0
langchain-ibm==0.2.1
langchain-text-splitters==0.3.0
langsmith==0.1.129
lomond==0.3.3
multidict==6.1.0
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pandas==2.1.4
proto-plus==1.24.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.31.0
yarl==1.13.1
zipp==3.20.2
bs4
